give the php files some much-needed TLC
