+++ package.json
@@ -73,14 +73,14 @@
     "@types/lodash": "^4.14.119",
     "@types/mini-css-extract-plugin": "^0.2.0",
     "@types/node": "^10.12.18",
-    "@types/react": "^16.7.18",
+    "@types/react": "^16.7.20",
     "@types/react-autosuggest": "^9.3.6",
     "@types/react-dom": "^16.0.11",
     "@types/react-test-renderer": "^16.0.3",
     "@types/rimraf": "^2.0.2",
     "@types/string-hash": "^1.1.1",
     "@types/uuid": "^3.4.4",
-    "@types/webpack": "^4.4.22",
+    "@types/webpack": "^4.4.23",
     "@wordpress/babel-plugin-makepot": "^2.1.3",
     "@wordpress/babel-preset-default": "^3.0.2",
     "@wordpress/browserslist-config": "^2.2.3",
@@ -111,7 +111,7 @@
     "sass-loader": "^7.1.0",
     "snapshot-diff": "^0.4.2",
     "style-loader": "^0.23.1",
-    "stylelint": "^9.9.0",
+    "stylelint": "^9.10.0",
     "stylelint-config-recommended-scss": "^3.2.0",
     "stylelint-scss": "^3.5.1",
     "ts-jest": "^23.10.5",
