--- src/php/views/reference-list.php
@@ -1,18 +1,0 @@
-<?php
-/**
- * Legacy reference list view.
- *
- * @package ABT
- */
-
-?>
-<div id='abt-reflist__root'>
-	<div id="abt_preload_container"></div>
-</div>
-<script type="text/javascript">
-	( function ( global ) {
-		var el = document.createElement( 'span' ); el.id = 'abt_changelog';
-		document.querySelector( '#abt-reflist > h2' ).appendChild( el );
-		global.HW_config={ selector:'#abt_changelog', account:'LJ4gE7' };
-	} )( window );
-</script>
